backend/recommandation-service/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Qdrant configuration
QDRANT_URL = os.getenv("QDRANT_URL", "http://localhost:6333")
QDRANT_COLLECTION = os.getenv("QDRANT_COLLECTION", "products")

# OpenAI configuration
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
OPENAI_EMBED_MODEL = os.getenv("OPENAI_EMBED_MODEL", "text-embedding-3-small")

# Service configuration
MAX_RECOMMENDATIONS = int(os.getenv("MAX_RECOMMENDATIONS", "10"))
DEFAULT_RECOMMENDATIONS = int(os.getenv("DEFAULT_RECOMMENDATIONS", "5"))

# Webhook configuration
WEBHOOK_SECRET = os.getenv("WEBHOOK_SECRET", "")
